# ENH(WiP): perspective API improvement for GroupClusterThreshold