… mvpa2/misc/neighborhood.py
… mvpa2/testing/tools.py
… mvpa2/tests/test_labeler.py
… mvpa2/tests/test_testing.py
… mvpa2/tests/test_usecases.py