to support other metrics, QE, etc

wanted to see what parameters should come into play etc.  And it feels that may be there should be some class hierarchy since many parameters are specific for cluster forming (threshold, qe) and require the two-stage process (first estimate thr_map, and only then clusters for the metric).  In that two stage process also -- may be  we need change/parameter to remove a possible bias due to using the same bootstraps/data twice (I don't have a clear/strong feeling on that yet).

If we agree on basic idea/API -- I could prep the functionality and see how it goes from there (either requires classes etc)