@nno 2016-01-17
Would this allow for TFCE and surface-based clustering?
As I wrote before, in my experience TFCE is much easier to implement if relevant cluster statistics are stored for each *feature* rather than each *cluster*. fixed-threshold clustering becomes then one special case.