@yarikoptic 2016-01-26
@nno -- added `map_postproc` parameter to the beast, so we could feed it with any callable which transforms original or bootstrapped sample anyhow.  Implementation details yet to be figured out.  First I will implement clustering using QE (yet to comprehend what was your concern)