@yarikoptic 2016-01-26
@PyMVPA/developers I have pushed my tentative RF to provide "clustering" using QEs.  The only additional somewhat demand is really that the dataset with permutations should carry a FlattenMapper somewhere with non-degenerate (i.e. non-None) space i.e. so we could identify which 'fa' to use for determining neighbors.  But that is only in the case if no explicit 'Labeler' (new construct implementing depth-first-search to identify clusters) was provided.  If provided - should be used.  Ideally we should get a unittest with data on surface.

I have disabled for now collection of 'spatial' statistics for clusters since that would require RFing of returned values anyways -- e.g. how should it look in case of surface dataset?  what if volumetric + temporal searchlight?  I thought that we should probably collect those using dataset's .fa and  query engine's arguments, but again not sure if those would always be consistent across all query engines.  IMHO such operation might better be performed explicitly after per each use case, e.g. `collect_spatial_clusters_stats(ds, fas=['voxel_indices'])`.  What do you think?

Also one test will break ATM since I was not sure on best way to address RFing of the internal `_get_map_cluster_sizes` which previously was taking a map and then calling .label() assuming pure Nd array to find neighbors in.  I could e.g. RF it so if no labels/num given it would flatten given data with our FlattenMapper and then proceed but imho it wouldn't be kosher.  I guess ideally I should provide it all those values so to test only that single invocation to count actual clusters ;)

What do you think?
Now I want to switch to enable other FW estimations, which wouldn't even care about "cluster" level anything