@yarikoptic 2016-01-26
BTW @hanke , if you could check it out on some of your real results  -- that would be great.  It should provide identical results -- if not -- there is or was a bug ;)