@hanke 2016-03-04
Doing the tests now....

First comment: The documentation needs to mention that in input dataset now needs to have a feature attribute indicating the spatial association of features. Before a mapper was sufficient -- which is now obsolete. This likely implies a change in existing analysis code, as the output of a searchlight, for example, does not have something like `voxel_indices`.