@hanke 2016-03-04
@yarikoptic I will not keep a full analysis running to check for identical results. The estimated runtime of the cluster size calculation is 20 hours (up from ~10min; see #427 for performance estimates for this test analysis).

I will massively reduce the number of bootstrap samples and try again.