@hanke 2016-03-04
Ok, I am confused wrt performance. But anyway, here is a direct comparison of the original implementation and this PR without parallelization, and compartmentalization:

````
# ORIGINAL
# n_proc=1, n_blocks=1, n_bootstrap=5000
683.82s user 17.23s system 100% cpu 11:37.20 total

# THIS PR
# n_proc=1, n_blocks=1, n_bootstrap=5000
589.81s user 17.94s system 100% cpu 10:04.26 total
````

so performance looks all good in this case. However, the results differ. How much they differ is kinda hard to tell. The clusters in the new implementation are not sorted by size, so IDs cannot easily be matched, but it seems that the biggest clusters there is an approximate equivalent in size.

Looking at the results in terms of the location in the volume it seems that they are structurally the same. There are a few smaller clusters different between the two algorithms, but given the not-necessarily stellar number of bootstraps I consider this good enough.

I am in favor of replacing the current implementation with this more flexible one. But we should make sure to keep parallelization in mind. It could be that I screwed up somewhere, though....