@hanke 2016-03-04
Ah, and sorting the cluster IDs again will be much appreciated!