@hanke 2016-03-04
FTR: Before a merge the docstring needs a full pass to adjust to the new implementation.