@yarikoptic 2016-03-11
@hanke  sorting by size should be inplace as of 9c7a3ff7063deb8b015bff9af40e928906878ba2 , note that differences still might happen if clusters have equal number of members.  Otherwise as of that commit I have refactored quite a few pieces primarily hopefully without change of functionality.  But now all 3 metrics for cluster level (size, sizenon0, maxsize) should work although still undertested I would say.